analysis_api/requirements.txt
numpy
pandas
matplotlib
flask
flask_cors
pillow
scipy
fooof
nltk
python-dotenv